+++ tests/Functional/ClientConfigurationProvider/BasicAuthTest.php
@@ -28,7 +28,7 @@ class BasicAuthTest extends ServerBasedTest
             })
         ;
 
-        $scrawler = new Scrawler($config, __DIR__);
+        $scrawler = new Scrawler($config, __DIR__ . '/output');
         $scrawler->run();
 
         $this->assertEquals('OK', InMemoryResultWriter::$results[0]['match']);