+++ tests/Functional/ResultWriterFilenameProvider/EntityPropertyFilenameProviderTest.php
@@ -34,7 +34,7 @@ class EntityPropertyFilenameProviderTest extends ServerBasedTest
             })
         ;
 
-        $scrawler = new Scrawler($config, __DIR__);
+        $scrawler = new Scrawler($config, __DIR__ . '/output');
         $scrawler->run();
 
         $files = glob(__DIR__ . '/output/test/*.json');